refactor: simplify browsergym hf sandbox example
